perf: move getDirtyValues util to avoid additional module perload
move around some files to avoid unecessary preload of assets
